+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 ## main
 
 ### ✨ Features and improvements
+- Support Terrain in Globe projection ([#4976](https://github.com/maplibre/maplibre-gl-js/pull/4976))
 - Improved performance of the `coveringTiles` (tile culling) function for globe ([#4937](https://github.com/maplibre/maplibre-gl-js/pull/4937))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
-- Fix `line` layer on Globe with Terrain Elevation
 - ⚠️ Fix level of detail at high pitch angle by changing which tiles to load ([#3983](https://github.com/maplibre/maplibre-gl-js/issues/3983))
 - _...Add new stuff here..._